feat: Implement plugin permission system  (#132)
* fix: Cleanup senders before stopping the network stack

* feat: Add plugin permission checks

+++ controller/src/application/plugin/runtime/wasm/ext/file.rs
@@ -1,7 +1,8 @@
-use anyhow::Result;
+use anyhow::{anyhow, Result};
 use tokio::fs::remove_dir_all;
 
 use crate::application::plugin::runtime::wasm::{
+    config::Permissions,
     generated::plugin::system::{
         self,
         types::{Directory, ErrorMessage},
@@ -11,6 +12,13 @@ use crate::application::plugin::runtime::wasm::{
 
 impl system::file::Host for PluginState {
     async fn remove_dir_all(&mut self, directory: Directory) -> Result<Result<(), ErrorMessage>> {
+        // Check if the plugin has permissions
+        if !self.permissions.contains(Permissions::ALLOW_REMOVE_DIR_ALL) {
+            return Err(anyhow!(
+                "Plugin tried to call remove_dir_all without the required permissions"
+            ));
+        }
+
         Ok(remove_dir_all(Self::get_directory(&self.name, &directory))
             .await
             .map_err(|error| format!("Failed to remove directory: {error}")))
